migration(micronauts): move packages to more common places.
