+++ src/main/java/org/terasology/web/model/artifactory/ArtifactRepository.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2015 MovingBlocks
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.terasology.web.model.artifactory;
+
+import java.io.IOException;
+import java.util.Collection;
+
+public interface ArtifactRepository {
+
+    public enum RepoType {
+        RELEASE,
+        SNAPSHOT
+    }
+
+    String getName();
+
+    RepoType getType();
+
+    Collection<String> getModuleNames();
+
+    void updateModule(String moduleName) throws IOException;
+
+    Collection<? extends ArtifactInfo> getModuleArtifacts(String moduleName);
+}
