+++ src/main/java/org/terasology/web/model/server/Result.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2015 MovingBlocks
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.terasology.web.model.server;
+
+public final class Result {
+
+    private final String message;
+    private final boolean success;
+
+    private Result(boolean success, String message) {
+        this.message = message;
+        this.success = success;
+    }
+
+    public static Result fail(String message) {
+        return new Result(false, message);
+    }
+
+    public static Result success(String message) {
+        return new Result(true, message);
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public boolean isSuccess() {
+        return success;
+    }
+}
