+++ src/main/java/org/terasology/web/model/web/ServerForm.java
@@ -1,0 +1,61 @@
+package org.terasology.web.model.web;
+
+/**
+ * Pojo for Server add/edit form.
+ */
+public class ServerForm {
+    private String name;
+    private String address;
+    private int port;
+    private String owner;
+    private String activeOn;
+    private String secret;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public int getPort() {
+        return port;
+    }
+
+    public void setPort(int port) {
+        this.port = port;
+    }
+
+    public String getOwner() {
+        return owner;
+    }
+
+    public void setOwner(String owner) {
+        this.owner = owner;
+    }
+
+    public String getActiveOn() {
+        return activeOn;
+    }
+
+    public void setActiveOn(String activeOn) {
+        this.activeOn = activeOn;
+    }
+
+    public String getSecret() {
+        return secret;
+    }
+
+    public void setSecret(String secret) {
+        this.secret = secret;
+    }
+}
