--- src/main/java/org/terasology/web/servlet/AboutController.java
@@ -1,45 +1,0 @@
-/*
- * Copyright 2015 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.terasology.web.servlet;
-
-import io.micronaut.core.util.CollectionUtils;
-import io.micronaut.http.HttpResponse;
-import io.micronaut.http.MediaType;
-import io.micronaut.http.annotation.Controller;
-import io.micronaut.http.annotation.Get;
-import io.micronaut.http.annotation.Produces;
-import io.micronaut.views.View;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.terasology.web.version.VersionInfo;
-
-/**
- * Show the about html page.
- */
-@Controller()
-public class AboutController {
-
-    private static final Logger logger = LoggerFactory.getLogger(AboutController.class);
-
-    @View("about")
-    @Produces(MediaType.TEXT_HTML)
-    @Get(value = "home")
-    public HttpResponse about() {
-        logger.info("Requested about as HTML");
-        return HttpResponse.ok(CollectionUtils.mapOf("version", VersionInfo.getVersion()));
-    }
-}
